C++ clean-up and introduction of 2GiB max alloc limit to DirectByteSink. Changed ByteSequence size from long to int.

+++ core/src/main/java/io/questdb/std/bytes/ByteSequence.java
@@ -35,11 +35,11 @@ public interface ByteSequence {
      * @param index byte index
      * @return byte at index
      */
-    byte byteAt(long index);
+    byte byteAt(int index);  // TODO: Convert to take `long` instead of `int`.
 
     /**
      * Number of bytes in the sequence.
      * Note that this is not called `length()` to avoid collision with `CharSequence` (and other) interfaces.
      */
-    long size();
+    int size();  // TODO: Convert this to return `long`. Also see `byte_sink.cpp` which restricts the size to `int`.
 }
